@@ -32,7 +32,7 @@
   ],
   "homepage": "https://github.com/speech-rule-engine/sre-l10n",
   "devDependencies": {
-    "@types/node": "17.0.14",
+    "@types/node": "20.2.5",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
     "@typescript-eslint/parser": "^5.7.0",
     "eslint": "8.8.0",
@@ -42,10 +42,10 @@
     "prettier": "^2.5.1",
     "terser-webpack-plugin": "5.3.1",
     "ts-loader": "9.2.6",
-    "typescript": "4.5.5",
+    "typescript": "5.0.2",
     "typescript-tools": "0.3.1",
-    "webpack": "5.68.0",
-    "webpack-cli": "4.9.2"
+    "webpack": "5.76.3",
+    "webpack-cli": "^5.0.1"
   },
   "files": [
     "LICENSE",
@@ -55,7 +55,8 @@
   "main": "dist/srel10n.js",
   "dependencies": {
     "wicked-good-xpath": "^1.3.0",
-    "xmldom-sre": "^0.1.31",
+    "@xmldom/xmldom": "0.9.0-beta.8",
     "yamljs": "^0.3.0"
-  }
+  },
+  "type": "module"
 }
